+++ src/models/Config.php
@@ -52,7 +52,7 @@ class Config extends Model
      * @var array HTML attribute options.
      */
     public array $attr = [
-        'class' => 'adWizard-ad',
+        'class' => 'adWizard',
         'style' => 'cursor:pointer',
     ];